store latest key on cosesigner (#55)
This paves the way for instantiating a cosesigner each
time we want to sign something. The reasoning is that it
lets us always get the latest key and have access to the key
name and version used. This lets us use the new KeyIdentifier()
method to set the Cose Sign1 headers usefully.

re: AB#9185

Co-authored-by: Henry Jewell <henry.jewell@datatrails.ai>